:recycle: Restructure of status resource
