+++ dataservice/api/status/README.md
@@ -1,0 +1,1 @@
+Contains version and health information about the Dataservice API.
