+++ dataservice/api/status/__init__.py
@@ -1,0 +1,1 @@
+from .resources import status_api
